moved opioid valueset generator testcases
