+++ src/main/java/org/opencds/cqf/tooling/terminology/opioidValuesetGenerator/testcase/generator/Helper.java
@@ -1,20 +1,11 @@
-package org.opencds.cqf.testcase.generator;
+package org.opencds.cqf.tooling.terminology.opioidValuesetGenerator.testcase.generator;
 
-import ca.uhn.fhir.context.FhirContext;
-import ca.uhn.fhir.parser.XmlParser;
-import com.google.gson.*;
 import org.apache.poi.ss.usermodel.Cell;
 import org.apache.poi.ss.usermodel.Row;
 import org.apache.poi.ss.usermodel.Sheet;
 import org.apache.poi.ss.usermodel.Workbook;
-import org.hl7.fhir.Parameters;
-import org.hl7.fhir.ParametersParameter;
-import org.opencds.cqf.Operation;
-import org.opencds.cqf.testcase.TestCaseOld;
-import org.opencds.cqf.testcase.TestCaseContainer;
+import org.opencds.cqf.tooling.terminology.opioidValuesetGenerator.testcase.TestCaseOld;
 
-import java.io.FileOutputStream;
-import java.io.IOException;
 import java.util.*;
